Array: Fixed integer overflow in low_lookup()
Thanks to Cezary Cerekwicki <ccerekwicki@opera.com> for the report.

+++ src/array.c
@@ -600,7 +600,7 @@ static INT32 low_lookup(struct array *v,
   b=v->size;
   while(b > a)
   {
-    c=(a+b)/2;
+    c=((unsigned INT32)a+(unsigned INT32)b)/2;
     q=fun(ITEM(v)+c,s);
     
     if(q < 0)
